+++ integtest/spec/helper/book.rb
@@ -1,6 +1,9 @@
 # frozen_string_literal: true
 
+require_relative 'book_conf'
+
 class Book
+  include BookConf
   attr_reader :title, :prefix
 
   ##
@@ -33,6 +36,17 @@ class Book
   # Should this book built directly to html (true) or to docbook first (false).
   attr_accessor :direct_html
 
+  ##
+  # Should the book declare itself noindex? Defaults to false.
+  attr_accessor :noindex
+
+  ##
+  # List of branches that are considered "live" for a book. Branches that are
+  # not live will be marked as `noindex`. Defaults to nil, meaning don't emit
+  # the list of live branches. In that case the docs build will default to
+  # *all* branches being live.
+  attr_accessor :live_branches
+
   def initialize(title, prefix)
     @title = title
     @prefix = prefix
@@ -40,10 +54,10 @@ class Book
     @sources = []
     @branches = ['master']
     @current_branch = 'master'
-    @respect_edit_url_overrides = false
     @lang = 'en'
-    @suppress_migration_warnings = false
-    @direct_html = false
+    @respect_edit_url_overrides = @suppress_migration_warnings = false
+    @direct_html = @noindex = false
+    @live_branches = nil
   end
 
   ##
@@ -69,37 +83,6 @@ class Book
   end
 
   ##
-  # The configuration needed to build the book.
-  def conf
-    # We can't use to_yaml here because it emits yaml 1.2 but the docs build
-    # only supports 1.0.....
-    conf = standard_conf
-    conf += "respect_edit_url_overrides: true\n" if @respect_edit_url_overrides
-    if @suppress_migration_warnings
-      conf += "suppress_migration_warnings: #{@suppress_migration_warnings}\n"
-    end
-    conf += <<~YAML
-      sources:
-      #{sources_conf}
-    YAML
-    indent conf, '    '
-  end
-
-  def standard_conf
-    <<~YAML
-      title:      #{@title}
-      prefix:     #{@prefix}
-      current:    #{@current_branch}
-      branches:   [ #{@branches.join ', '} ]
-      index:      #{@index}
-      tags:       test tag
-      subject:    Test
-      lang:       #{@lang}
-      direct_html: #{@direct_html}
-    YAML
-  end
-
-  ##
   # The html for a link to a particular branch of this book.
   def link_to(branch)
     url = "#{@prefix}/#{branch}/index.html"
@@ -108,46 +91,5 @@ class Book
       decoration = " [#{@current_branch}]"
     end
     %(<a class="ulink" href="#{url}" target="_top">#{@title}#{decoration}</a>)
-  end
-
-  private
-
-  def sources_conf
-    yaml = ''
-    @sources.each do |config|
-      yaml += "\n-\n#{source_conf config}"
-    end
-    indent yaml, '  '
-  end
-
-  def source_conf(config)
-    yaml = <<~YAML
-      repo:    #{config[:repo]}
-      path:    #{config[:path]}
-    YAML
-    yaml += alternatives_conf config[:alternatives]
-    yaml += "private: true\n" if config[:is_private]
-    yaml += map_branches_conf config[:map_branches]
-    indent yaml, '  '
-  end
-
-  def alternatives_conf(conf)
-    return '' unless conf
-
-    yaml = ''
-    yaml += "alternatives:\n"
-    yaml += "  source_lang: #{conf[:source_lang]}\n"
-    yaml += "  alternative_lang: #{conf[:alternative_lang]}\n"
-    yaml
-  end
-
-  def map_branches_conf(map_branches)
-    return '' unless map_branches
-
-    yaml = "map_branches:\n"
-    map_branches.each_pair do |key, value|
-      yaml += "  #{key}: #{value}\n"
-    end
-    yaml
   end
 end
